Remove link tag to CSS that can't be loaded
"css/bootstrap.min.css" can't be loaded as we're running from a local
file.
This isn't a problem because we have the CDN link that does work further
down.
Removing the broken line, so that it doesn't show us in debug log, to
make it obvious that it's not the problem when something else goes
wrong.